hotfix: 화면이동 수정
