+++ app/src/main/AndroidManifest.xml
@@ -37,13 +37,11 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         
-<<<<<<< HEAD
         <!-- Google Play Services 메타데이터 -->
         <meta-data
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
         
-=======
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="${applicationId}.fileprovider"
@@ -54,7 +52,6 @@
                 android:resource="@xml/file_paths" />
         </provider>
 
->>>>>>> main
         <activity
             android:name=".SplashActivity"
             android:exported="false" />
