+++ app/src/main/java/com/example/clo/ProfileSettingsActivity.kt
@@ -124,7 +124,7 @@ class ProfileSettingsActivity : AppCompatActivity() {
                     Toast.makeText(this, "설정에서 검색 클릭", Toast.LENGTH_SHORT).show()
                     true
                 }
-                R.id.navigation_home -> {
+                R.id.menu_home -> {
                     val intent = Intent(this, MainActivity::class.java)
                     startActivity(intent)
                     finish()
