+++ app/src/main/res/drawable/button_border_black_less_round.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@android:color/white" />
+    <stroke android:width="2dp" android:color="@android:color/black" />
+    <corners android:radius="8dp" />
+</shape>
