+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,22 +3,22 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/navigation_home">
+    app:startDestination="@id/menu_home">
 
     <fragment
-        android:id="@+id/navigation_home"
+        android:id="@+id/menu_home"
         android:name="com.example.clo.HomeFragment"
         android:label="Home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/navigation_search"
+        android:id="@+id/menu_search"
         android:name="com.example.clo.SearchFragment"
         android:label="Search"
         tools:layout="@layout/fragment_search" />
 
     <fragment
-        android:id="@+id/navigation_profile"
+        android:id="@+id/menu_mypage"
         android:name="com.example.clo.MyPageFragment"
         android:label="MyPage"
         tools:layout="@layout/fragment_my_page" />
