+++ app/src/main/res/xml/file_paths.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<paths xmlns:android="http://schemas.android.com/apk/res/android">
+    <external-path name="external_files" path="."/>
+    <external-cache-path name="external_cache" path="."/>
+    <cache-path name="cache" path="."/>
+    <files-path name="files" path="."/>
+</paths>
